@@ -1,5 +1,13 @@
 module github.com/yi-jiayu/secure
 
-go 1.15
+go 1.20
 
 require gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
+
+require (
+	github.com/miekg/dns v1.1.55 // indirect
+	golang.org/x/mod v0.7.0 // indirect
+	golang.org/x/net v0.2.0 // indirect
+	golang.org/x/sys v0.2.0 // indirect
+	golang.org/x/tools v0.3.0 // indirect
+)
